Refactor: Split mapping and query test suites
